refactor storage backends fixtures centrally
